// cypress/cypress/support/helpers/enums/values-for-assertions.ts
export enum assertionValues {
  enabled = 'enabled',
  disabled = 'disabled',
  exist = 'exist',
  notExist = 'not.exist'
}

export enum paymentTitles {
  BDO = 'BDO Unibank Inc.',
  union = 'Union Bank',
  landBank = 'Land Bank of the Philippines',
  philippineIslands = 'Bank of the Philippine Islands',
  developmentBank = 'Development Bank of the Philippines',
  RCBC = 'RCBC',
  metrobank = 'Metrobank',
  chinaBank = 'China Banking Corporation',
  PNB = 'Philippine National Bank',
  securityBank = 'Security Bank',
  robinsons = 'Robinsons Bank',
  PSBank = 'PS Bank',
  GCash = 'GCash',
  Maya = 'Maya'
}